throw error instead of overwriting builtin export (#1225)
